Add files via upload
Resulting file from disease_associated2gene.py. The output contains the name of the gene, the disease associated and if it is X-linked recessive or dominant.

This output is going to be used to filter the massive vcf annotated file that has been obtained from Genome Project, in order to select exclusively the variants related to those genes.